requirements.txt
pygame==2.5.2
grpcio==1.51.3
grpcio-tools==1.51.3
